+++ test/sql/halfvec_functions.sql
@@ -6,6 +6,9 @@ SELECT '[1,2,3]'::halfvec * '[4,5,6]';
 SELECT '[65519]'::halfvec * '[65519]';
 SELECT '[1e-7]'::halfvec * '[1e-7]';
 
+SELECT '[1,2,3]'::halfvec || '[4,5]'::halfvec;
+SELECT array_fill(0, ARRAY[16000])::halfvec || '[1]'::halfvec;
+
 SELECT '[1,2,3]'::halfvec < '[1,2,3]';
 SELECT '[1,2,3]'::halfvec < '[1,2]';
 SELECT '[1,2,3]'::halfvec <= '[1,2,3]';